Tcl scripts should use tclsh found by configure
The configure script searches for tclsh (and fails if it can't be
found).

This change touches up the makefile and the src/*tcl scripts so that
their `#!` lines use the tclsh that was found at build time rather than
trusting what's found in the environment at run time.

This change is useful for tools like
[spack](https://github.com/llnl/spack), which very carefully manage and
provide prerequisites at build time so that users do not need to be
aware of them at run time.