Adding additional explanatory text about LMOD_AVAIL_STYLE
I added a brief explanatory text motivating use of the grouped
labeling.

At the end of the document, I added some text from

https://sourceforge.net/p/lmod/mailman/message/32759572/

where you had a clear and more complete explanation of the use
of MODULE_AVAIL_STYLE.

We are looking at EasyBuild, but we are using our current Lmod
installation, for which we export MODULE_AVAIL_STYLE=grouped,
and while we are looking at EB, it's more useful to change back
to the system style.  I didn't find anything to revert in the
Lmod documentation, and this is the result.