Add ability to test a shell name on "module-info shell"
Update Tcl "module-info shell" command to be able to test a shell name
argument and return if this shell name corresponds to the current shell
of the user.

This change helps to get the same behavior than Environment Modules
(introduced in version 4.0) and makes Lmod compatible with the Tcl
modulefile produced by the Singularity HPC tool.