Add "report" support for Tcl modules
The "report" modulefile command is available in Environment Modules to
provide a way for modulefile developer to print text to user, like
achieved on Lua modulefiles with the "LmodMessage()" function.

This commit adds "report" support for Tcl modulefiles and make it map to
the "LmodMessage()" Lua function when translated by tcl2lua.tcl.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>